resources/test/tests/idlharness/IdlInterface/test_exposed_attribute.html

<!DOCTYPE HTML>
<html>

<head>
  <title>IdlInterface Exposed attributes</title>
</head>

<body>
  <div id="log"></div>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/WebIDLParser.js"></script>
  <script src="/resources/idlharness.js"></script>
  <script src="../helper.js"></script>
  <script>
    "use strict";

    test(function () {
      const idlArray = new IdlArray();
      idlArray.add_idls(`
          [Exposed=(Window),Exposed=(Worker)]
          interface Document {};
      `);
      idlArray.assert_throws(
          'Multiple \'Exposed\' extended attributes on Document',
          i => i.test());
    }, 'should throw for multiple exposed attributes.');

    test(function () {
      const idlArray = new IdlArray();
      idlArray.add_idls(`
        [Exposed=(Window,Worker)]
        interface Document {};

        [Exposed=(Window)]
        partial interface Document {
          readonly attribute boolean Foo;
        };
      `);
      idlArray.test();
    }, 'should not throw for exposed attribute on partial.');
  </script>
</body>

</html>
